// src/modules/user/infraestructure/model/user.model.ts
export interface UserModel {
    uuid: string;
    national_id: string;
    name: string;
    f_lastname: string;
    s_lastname: string | null;
    dob: string;
    phone: string | null;
    email: string | null;
    access_level: number | null;
    headquarter: string;
    council_role: string | null;
}